Move MCP connect() lifecycle into agent/mason/mcp_runtime.py
build_mcp_servers() (the user knob) stays in agent/mcps.py; the connect()/MCPServerManager
plumbing moves to mason/ with the other SDK-migration candidates. Also drops a dead json
import and applies safe ruff fixes to the moved files.
